Fix regex that plucked last week issue's number
We were getting a 404 fetching last week issue, because the url was malformed. We use a regex to extract the identifier of the issue to be downloaded. Such regex assumed that the identifier was always made up of numbers. But this is not always true. So, I made the regx be more permissive.